AuthRequest applied on Payment Method Resource

+++ src/payment-method/payment-method.service.spec.ts
@@ -10,6 +10,7 @@ import { PaymentMethodService } from '@payment_method/payment-method.service';
 
 /* Entity */
 import { PaymentMethod } from '@payment_method/entities/payment-method.entity';
+import { User } from '@user/entities/user.entity';
 
 /* DTO's */
 import { UpdatePaymentMethodDto } from '@payment_method/dto/update-payment-method.dto';
@@ -110,7 +111,7 @@ describe('PaymentMethodService', () => {
       } catch (error) {
         expect(error).toBeInstanceOf(NotFoundException);
         expect(error.message).toBe(
-          `The Payment Method with id: ${id} not found`,
+          `The Payment Method with ID: ${id} not found`,
         );
       }
     });
@@ -120,7 +121,7 @@ describe('PaymentMethodService', () => {
       jest.spyOn(repository, 'findOne').mockResolvedValue(null);
 
       await expect(service.findOne(id)).rejects.toThrowError(
-        new NotFoundException(`The Payment Method with id: ${id} not found`),
+        new NotFoundException(`The Payment Method with ID: ${id} not found`),
       );
     });
   });
@@ -128,23 +129,23 @@ describe('PaymentMethodService', () => {
   describe('create payment methods services', () => {
     it('create should return a Payment Method', async () => {
       const paymentMethod = generatePaymentMethod();
-
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'create').mockReturnValue(paymentMethod);
       jest.spyOn(repository, 'save').mockResolvedValue(paymentMethod);
 
-      const { statusCode, data } = await service.create(paymentMethod);
+      const { statusCode, data } = await service.create(paymentMethod, userId);
       expect(statusCode).toBe(201);
       expect(data).toEqual(paymentMethod);
     });
 
     it('create should return Conflict Exception when name Payment Method exists', async () => {
       const paymentMethod = generatePaymentMethod();
-
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'create').mockReturnValue(paymentMethod);
       jest.spyOn(repository, 'save').mockResolvedValue(paymentMethod);
 
       try {
-        await service.create(paymentMethod);
+        await service.create(paymentMethod, userId);
       } catch (error) {
         expect(error).toBeInstanceOf(ConflictException);
         expect(error.message).toBe(
@@ -158,6 +159,7 @@ describe('PaymentMethodService', () => {
     it('update should return message: have been modified', async () => {
       const paymentMethod = generatePaymentMethod();
       const id = paymentMethod.id;
+      const userId: User['id'] = 1;
       const changes: UpdatePaymentMethodDto = { name: 'newName' };
 
       jest.spyOn(repository, 'findOne').mockResolvedValue(paymentMethod);
@@ -166,26 +168,27 @@ describe('PaymentMethodService', () => {
         .mockReturnValue({ ...paymentMethod, ...changes });
       jest.spyOn(repository, 'save').mockResolvedValue(paymentMethod);
 
-      const { statusCode, message } = await service.update(id, changes);
+      const { statusCode, message } = await service.update(id, userId, changes);
       expect(repository.findOne).toHaveBeenCalledTimes(1);
       expect(repository.merge).toHaveBeenCalledTimes(1);
       expect(repository.save).toHaveBeenCalledTimes(1);
       expect(statusCode).toBe(200);
       expect(message).toEqual(
-        `The Payment Method with id: ${id} has been modified`,
+        `The Payment Method with ID: ${id} has been modified`,
       );
     });
 
     it('update should throw NotFoundException if Payment Method does not exist', async () => {
       const id = 1;
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'findOne').mockResolvedValue(null);
 
       try {
-        await service.update(id, { name: 'newName' });
+        await service.update(id, userId, { name: 'newName' });
       } catch (error) {
         expect(error).toBeInstanceOf(NotFoundException);
         expect(error.message).toBe(
-          `The Payment Method with id: ${id} not found`,
+          `The Payment Method with ID: ${id} not found`,
         );
       }
     });
@@ -195,26 +198,27 @@ describe('PaymentMethodService', () => {
     it('remove should return status and message', async () => {
       const paymentMethod = generatePaymentMethod();
       const id = paymentMethod.id;
-
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'findOne').mockResolvedValue(paymentMethod);
       jest
         .spyOn(repository, 'merge')
         .mockReturnValue({ ...paymentMethod, isDeleted: true });
       jest.spyOn(repository, 'save').mockResolvedValue(paymentMethod);
 
-      const { statusCode, message } = await service.remove(id);
+      const { statusCode, message } = await service.remove(id, userId);
       expect(statusCode).toBe(200);
       expect(message).toEqual(
-        `The Payment Method with id: ${id} has been deleted`,
+        `The Payment Method with ID: ${id} has been deleted`,
       );
     });
 
     it('remove should throw NotFoundException if Payment Method does not exist with Rejects', async () => {
       const id = 1;
+      const userId: User['id'] = 1;
       jest.spyOn(repository, 'findOne').mockResolvedValue(null);
 
-      await expect(service.remove(id)).rejects.toThrowError(
-        new NotFoundException(`The Payment Method with id: ${id} not found`),
+      await expect(service.remove(id, userId)).rejects.toThrowError(
+        new NotFoundException(`The Payment Method with ID: ${id} not found`),
       );
     });
   });
